Send BCC URL to the Labber machine along with Scenario
> Warning: A breaking API change

> Description
-------------
This change allows the Labber machine to dynamically response,
with a produced logfile, to the right BCC machine. Before,
the Labber machine was statically configured in this respect.

This is change is tighly coupled with related modifications in
the 'tergite-labber-connector' repository. As such it is a
breaking change, if the Labber machine is not updated accordingly.

> Verification
--------------
Functionality has been verified with the whole stack:
Qiskit, MSS, BCC, Labber.