Step to Python 3.8
> Description
-------------
Tergite BCC has been verified to work with Python 3.8. Minimal
versions of dependent packages have been stepped inline with this
verification.

The setup.py triggers installation of requirements. The BCC
is expected to be installed in a development mode with

    pip install -e .

exactly as Tergite MSS.

> Verification
--------------
Functionality has been verified with the whole stack:
Qiskit, MSS, BCC, Labber.