[BC-20] Enforce job_id to be RFC 4122 UUID4 compliant
Description
-----------
This commit enforces the job_id to be RFC 4122 UUID4 compliant.
The validation is performed at the rest_api module, right after
a job is received. There is a related commit in the Tergite-Labber
Connector.

Additional benefits
-------------------
This enforcement allows us to use the job_id in a filename
for jobs, scenarios and logfiles, without the need to resort
to additional unique identifiers. This makes the whole
processing chain more streamlined: especially until we
have a job supervisor in place.

Verification
------------
The whole processing chain has been tested: Labber, BCC, MSS,
Qiskit. Installation of all involved packeges was done from
scratch.