[BC-50] Submit calibration as a job, big change set
This commit is quite big, and it implements the main task of BC-50,
that the calibration supervisor submits a calibration as a job to the
execution_worker instead of, as previosly, sending a scenario directly
to Labber.

`calibration_graph_parser`:

Replace `check_f` and `cal_f` with `calibration_fn` and `check_fn`,
respectively `dummy_check.py` and `dummy_cal.py` in the
`cal_routines/` directory.

These generate jobs instead of scenarios. As a first approximation,
this file contains definitions of calibration functions and check
functions that are mapped from the corresponding nodes in the
calibration graph, but all the check functions point to the same
check function, and all the calibration functions point to the same
calibration function. In the end, the calibration function calls the
signal demodulation job, and so does the check function, but with a
smaller number of points in the measurment. This is just to show
that one can put functions here that are determined by the nodes
in the calibration graph.

`calibration_supervisor`:

* Obvious import changes
* Change `REST_API_MAP` for jobs instead of scenarios
* Also remove unnecassary "calibrated" flag from old code
* Change `MEASUREMENT_ROUTINES` to `MEASUREMENT_JOBS` and point to the
  new calibration mockup functions

* In `check_data`:

* Send job request to `execution_worker` instead of sending a
   scenario to Labber.
* Fetch `job_id` from redis, from postprocessing, instead of
  `shots`.

* In `calibrate`:

* Send job instead of scenario
* Fetch postprocessed `results:job_id` instead of
  `results:shots` from Redis

* Define `send_job`:

* Send the job dictionary to `execution_worker` instead of sending a
  scenario to Labber.
* Use `asyncio.Event` to await replies from postprocessing, via message server

* `execution_worker.py`:
* In `job_execute`
* In case the job comes from the calibration supervisor, set the
     scenario tags' third slot to indicate it

* `postprocessing_worker.py`:

* In `logfile_postprocess`:
* Extract `is_calibration_sup_job` from scenario tags
* Take care of the case `"demodulation_scenario"` in case 1 above
* Also return is_calibration_sup_job flag to indicate whether or
  not we have a calibration supervisor job
* Replace `process_dummy` with `process_demodulation` postprocessing
  method
* Extract `is_calibration_sup_job` from tags
* Set "`results:job_id`" to `job_id` in Redis, as a placeholder
  for data extracted from logfile

* In `postprocessing_success_callback`:
* Use `is_calibration_sup_job` to indicate calibration supervisor
  job instead of prevoius `"calibration"` string
* In the notification message to the calibration supervisor, have a
  tag "job_done:" as prefix.