[BC-81] Move measurement job functions to a directory of its own
The helper functions to create measurement jobs are in principle not
calibration framework specific. This change moves the measurement job
funtions to an own directory, also to be populated with config files in
a future commit. Then one can choos to submit such a job as called
from the calibration framework, or as a stand-alone experiment.