[BC-81] Refactor scenario population, apply on pulsed resonator spectroscopy
Instead of having all the scenario parameters in the job dict, we put
the fixed parameters in a default file (in TOML format), and let the
variable parameters be in the job 'params' field. The
scenario_scripts.py loads the defaults from the TOML file, and then
the parameters provided from the job overrides them. The parameters
specified in the mk_job_ header are mandatory, and then the user can
provide more parameters to override the defaults from the TOML file.

We also let the user control, as an option, the name of the job, and
the flag that indicates whether or not the job emanates from the
calibration supervisor.