[BC-81] Rename parameter ranges more consistently
We change FOO_{start|stop}_BAR into FOO_BAR_{start|stop}, where it applies.
More precisely, we do a regex query-replace (emacs syntax)

\([a-z_]+\)_\(start\|stop\)_\([a-z]+\) → \1_\3_\2

so that foo_bar is not split up.

For instance, readout_power is one thing, and then _start and _stop
are suffixes to denote endpoints of the interval.