[BC-87] Add a subset of job dict's keys to job_supervisor's entry
Add these keys when the job enters the pipeline: e.g.,
registration_worker.  Subsequent processes that needs the job name,
the tag indicating whether or not the job is requested from the
calibration_supervisor, can be retrieved from the job_supervisor
without having to rely on Labber logfile tags, for instance.