[BC-87] Call the Labber.LogFile() constructor at most once for each job
All the Labber specific helper functions use the Labber.LogFile class
rather than a filepath. When a logfile arrives for post-processing,
the first access to the contents is the extraction of the job_id. This
requires the Labber.LogFile() constructor to be invoked. This takes
time in the order of a second, and therefore we should avoid calling
it more than once for the same job. The LabberParsing function in
qtanalysis repo has been changed, so it takes a Labber.LogFile object.