[BC-101] Rename measurement parameters to be more coherent
The readout drive and drive frequencies are renamed according to the pattern

FREQUENCY_if of the IF / low band part, and FREQUENCY_lo for the LO /
high band part.

Suffixes _range_type, _start, and _stop are added accordingly.

In scenario_scripts.py we also add a log message to see the scenario
that was created.