[BC-101] Change from index: int to component_id: str in storage.py
This makes it possible to have free naming of components, as long as
they are configured in the for device properties and property
initialization template (see backend_properties_config/).

We also rename and clarify the role of the
calibration/calibration_common.write_calibration_result function: that
let us choose whether to only store a result internally for the
calibration_supervisor or to also publish it to the "public" Redis
storage. This could have been a separate commit, but ...