[BC-110] Let the user specify a list of calibration goal nodes
The user can specify a comma-separated list of node names, they will
cause the graph generated to contain only the nodes that they depend
on. Example(see the new example graph):

CALIBRATION_GOALS="e,d" CALIBRATION_GRAPH=calibration_graphs/subgraph_test.json python calibration_graph_parser.py

This is useful when you have a big calibration graph, and want to
perform calibration on just a subset of it.

* Also:

      * Fix undefined prefix variable in calibration lib dummy functions

      * Remove unused "fidelity" tags in calibration_graph_parser

A note on the graph representation legacy code: There is a phase that
removes "redundant edges". However, that should already fall out
automatically from topological sorting, so it shouldn't be
needed. Also the topological sorting itself is not mentioned in the
Optimus paper (see doc/calibration.md), and there is no explanation
why it is there. Perhaps an optimization.