Update the 'result' property of job in redis with job results during postprocessing
The endpoint '/jobs/{job_id}/result seems to expect to extract the job results from' results of a job run
the 'result' property of the job dump in redis.
However, when all workers are checked, it seems like no worker at the time of
this change was actually setting this property.
The closest thing was in the 'postprocessing_success_callback' which instead of
setting the 'result' to the actual 'result', it was setting it to the job_id.

This was fixed, and also the setting of Location.FINAL_Q and Location.FINAL_W
added in the postprocessing_success_callback because, again, these locations
seemed to never be set anywhere else.