Fix bash error: 'line 42: [: =: unary operator expected'
This was caused by the fact that when variables are not set in bash,
they become empty space and thus a conditional like `[ $IS_SYSTEMD = "true" ]`
becomes `[ = "true" ]` if `IS_SYSTEMD` is not set.

It was solved by wrapping double-quotes on the variable so that if it
is not set, it collapses into an empty string i.e. `[ "$IS_SYSTEMD" = "true" ]`
becomes `[ "" = "true" ]` if `IS_SYSTEMD` is not set.