Run python modules with python -m in start_bcc.sh
When `conda activate <new-env>` is run on an already activated conda environment,
both conda environments are active. Running modules like 'uvicorn', 'pip',
'rq' without specifying which environment we want to be used can cause
the running module to select an environment we are not interested in.

Specifying `python -m <module>` ensures that the <module> is run using
the python of the current environment as seen in the terminal.