Fix error 'number of qubits is greater than max'  when single qubit simulator is used
It turns out that the coupling map required for a backend without couplers
is one where the qubits seem like they are coupled to themselves
e.g. [[0, 0]]. I added this special case when generating the coupling map from
coupling_dict.

Also, there are two things that can be confusing. A qubit's id is different
from its index or position in the list of available qubits. For example if
we have [q12, q13, q14], the index of q12 is 0 but its id is 12.
coupling_map always works with the index while the code generating cz gates
uses ids. Thus the mapping of qubit ids and the coupler id was added to the
response sent to tergite SDK