Remove storage_file lib
Why
---
StorageFile code was hard to understand and thus hard to maintain.

What was Done
---

- Replaced `StorageFile` with `QuantumJob` data transfer objects

In this pull request, we attempt to separate data from procedure.
Experiment data passes through the system as instances of the `QuantumJob` class.

Executors operate on instances of these `QuantumJob` data transfer objects.
These objects are shared between processes (or workers) by saving them as is in HDF5 files.
This means we can decide one day to swap out HDF5 file storage for anything else and all that
would needed would be similar functions but for the new medium.

- Added explicit types for most of the data

In order to ease reading the code, we have added dataclass, enum and pydantic model definitions
for the most of the data that is operated on by the executor and post processor workers.

- Replaced complex if-statements in qobj compilation with dict lookups

When converting PulseQobj instances to quantify-specific or qiskit-dynamics-specific
data, there were some hard-to-comprehend if-statements used in the past.
These have been swapped out for dictionary look-ups and new data type class definitions.
We now have class definitions for each possible instruction that can be run on either
quantify or qiskit dynamics.

- Renamed misleading function and method names

Some method/function names were misleading.
For instance `QuantumExecutor.run` and `QauntumExecutor.run_experiments` were confusing.
It turned out that `run_experiments` was the actual `run` method called by external code
and the old `run` method was supposed to be a private or protected method hidden from external
callers. It was also supposed to generate the quantify-specific or qiskit-dynamics-specific
experiment data. It was thus renamed `_to_native_experiments`.

There are a number of such renaming tasks done all over the code.

- Archived the old StorageFile code

There was some code in the StorageFile that no one currently in the team understood.
However, it seemed important. This was put in the `app/libs/quantum_executor/base/quantum_job/archive.py`.
In fact all code related to the original storage_file library was put in that module and commented out.
Those who will implement such things as new measurement levels and measurement returns can check that module.

- Removed redundant data transformations

The shape of the data in the HDF5 was originally the transposed version of the raw results
from the executors. However, reading the code, we realised that there was no need to do that transformation.
We thus removed it.

- Replaced for-loops with numpy vectorized functions where necessary

There were some places where we had originally had for-loops that were iterating through
numpy arrays or xarray datasets. We replaced these with the more performant vectorized
equivalents.

More room for improvement

- The DAG creation in quantify program was not touched. It needs a little touch up.
- The postprocessing and execution workers are still in the same state they were in when the project was a proof of concept.
- The `app/libs/properties` library needs some love. The biggest impact will be the `app/libs/properties/storage` module.
    - It would be nice to have that storage module less coupled to redis because it is possible that some people or users
      might prefer using other databases/caches e.g. SQLite or memcached etc.